+++ packages/angular/ssr/BUILD.bazel
@@ -44,6 +44,7 @@ ng_package(
     tags = ["release-package"],
     deps = [
         ":ssr",
+        "//packages/angular/ssr/node",
     ],
 )
