+++ packages/angular/ssr/node/public_api.ts
@@ -6,4 +6,8 @@
  * found in the LICENSE file at https://angular.dev/license
  */
 
-export * from './private_export';
+export {
+  CommonEngine,
+  type CommonEngineRenderOptions,
+  type CommonEngineOptions,
+} from './src/common-engine/common-engine';
