+++ packages/angular/ssr/node/src/common-engine/peformance-profiler.ts
