+++ packages/angular/ssr/private_export.ts
@@ -16,3 +16,5 @@ export {
   setAngularAppManifest as ɵsetAngularAppManifest,
   setAngularAppEngineManifest as ɵsetAngularAppEngineManifest,
 } from './src/manifest';
+
+export { InlineCriticalCssProcessor as ɵInlineCriticalCssProcessor } from './src/utils/inline-critical-css';
