+++ packages/angular_devkit/build_angular/src/builders/ssr-dev-server/specs/ssl_spec.ts
@@ -25,7 +25,7 @@ describe('Serve SSR Builder', () => {
       'src/main.server.ts': `
         import 'zone.js/node';
 
-        import { CommonEngine } from '@angular/ssr';
+        import { CommonEngine } from '@angular/ssr/node';
         import * as express from 'express';
         import { resolve, join } from 'node:path';
         import { AppServerModule } from './app/app.module.server';
